@@ -11,45 +11,61 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.10",
-    "@angular/cdk": "^7.3.5",
-    "@angular/common": "~7.1.0",
-    "@angular/compiler": "~7.1.0",
-    "@angular/core": "~7.1.0",
-    "@angular/fire": "^5.1.2",
-    "@angular/forms": "~7.1.0",
-    "@angular/material": "^7.3.5",
-    "@angular/platform-browser": "~7.1.0",
-    "@angular/platform-browser-dynamic": "~7.1.0",
-    "@angular/router": "~7.1.0",
+    "@agm/core": "^3.0.0-beta.0",
+    "@agm/snazzy-info-window": "^3.0.0-beta.0",
+    "@angular/animations": "^10.1.0",
+    "@angular/cdk": "^10.2.0",
+    "@angular/common": "~10.1.0",
+    "@angular/compiler": "~10.1.0",
+    "@angular/core": "~10.1.0",
+    "@angular/fire": "^6.0.2",
+    "@angular/forms": "~10.1.0",
+    "@angular/material": "^10.2.0",
+    "@angular/material-moment-adapter": "^10.2.0",
+    "@angular/platform-browser": "~10.1.0",
+    "@angular/platform-browser-dynamic": "~10.1.0",
+    "@angular/router": "~10.1.0",
+    "@firebase/app": "^0.6.10",
+    "@types/googlemaps": "^3.39.13",
+    "@types/moment": "^2.13.0",
+    "angular-chart.js": "^1.1.1",
     "angular2-text-mask": "^9.0.0",
-    "core-js": "^2.6.5",
-    "firebase": "^5.9.1",
+    "chart.js": "^2.9.3",
+    "core-js": "^3.6.5",
+    "file-saver": "^2.0.2",
+    "firebase": "^7.19.1",
     "hammerjs": "^2.0.8",
-    "rxjs": "~6.3.3",
+    "moment": "^2.27.0",
+    "ng2-charts": "^2.4.0",
+    "ngx-moment": "^5.0.0",
+    "npm-install-peers": "^1.2.1",
+    "rxjs": "~6.6.2",
+    "saturn-datepicker": "^8.0.5",
+    "snazzy-info-window": "^1.1.1",
     "text-mask-addons": "^3.8.0",
-    "tslib": "^1.9.0",
-    "zone.js": "^0.8.29"
+    "tslib": "^2.0.1",
+    "xlsx": "^0.16.6",
+    "zone.js": "^0.11.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.6",
-    "@angular/cli": "~7.1.4",
-    "@angular/compiler-cli": "^7.2.10",
-    "@angular/language-service": "~7.1.0",
-    "@types/jasmine": "^2.8.16",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.1.1",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^2.0.5",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.1.6"
+    "@angular-devkit/build-angular": "^0.1001.0",
+    "@angular/cli": "^10.1.0",
+    "@angular/compiler-cli": "^10.1.0",
+    "@angular/language-service": "~10.1.0",
+    "@types/jasmine": "^3.5.14",
+    "@types/jasminewd2": "~2.0.8",
+    "@types/node": "~14.6.4",
+    "codelyzer": "~6.0.0",
+    "jasmine-core": "~3.6.0",
+    "jasmine-spec-reporter": "~5.0.2",
+    "karma": "^5.2.1",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "^3.0.3",
+    "karma-jasmine": "~4.0.1",
+    "karma-jasmine-html-reporter": "^1.5.4",
+    "protractor": "~7.0.0",
+    "ts-node": "~9.0.0",
+    "tslint": "~6.1.3",
+    "typescript": "~4.0.2"
   }
 }
