@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { DomSanitizer } from '@angular/platform-browser';
-import { MatIconRegistry } from '@angular/material';
+import { MatIconRegistry } from '@angular/material/icon';
 
 @Component({
   selector: 'app-root',
@@ -21,11 +21,15 @@ export class AppComponent {
       .addSvgIcon('ic-products', sanitizer.bypassSecurityTrustResourceUrl('assets/icons/icon-productos-off.svg'))
       .addSvgIcon('ic-users', sanitizer.bypassSecurityTrustResourceUrl('assets/icons/icon-usuarios-off.svg'))
       .addSvgIcon('ic-clients', sanitizer.bypassSecurityTrustResourceUrl('assets/icons/icon-clientes-off.svg'))
+      .addSvgIcon('ic-back', sanitizer.bypassSecurityTrustResourceUrl('assets/icons/icon-vendedor-back-gray.svg'))
       .addSvgIcon('ic-before', sanitizer.bypassSecurityTrustResourceUrl('assets/icons/before.svg'))
       .addSvgIcon('ic-next', sanitizer.bypassSecurityTrustResourceUrl('assets/icons/next.svg'))
       .addSvgIcon('ic-edit', sanitizer.bypassSecurityTrustResourceUrl('assets/icons/icon-edit-gray.svg'))
       .addSvgIcon('ic-dropdown', sanitizer.bypassSecurityTrustResourceUrl('assets/icons/ic-icon-dropdown-orange.svg'))
-      .addSvgIcon('ic-close-circle', sanitizer.bypassSecurityTrustResourceUrl('assets/icons/close-circle.svg'));
+      .addSvgIcon('ic-close-circle', sanitizer.bypassSecurityTrustResourceUrl('assets/icons/close-circle.svg'))
+      .addSvgIcon('ic-close-grey', sanitizer.bypassSecurityTrustResourceUrl('assets/icons/icon-close-grey.svg'))
+      .addSvgIcon('ic-calendar', sanitizer.bypassSecurityTrustResourceUrl('assets/icons/icon-calendar-orange.svg'))
+      ;
 
   }
 }
