@@ -21,5 +21,29 @@ export const NUMBER_MASK = createNumberMask({
   requireDecimal: false,
   allowNegative: false,
   allowLeadingZeroes: false,
-  integerLimit: 3
+  integerLimit: 4
+});
+
+export const PERCENTAGE_MASK = createNumberMask({
+  prefix: '',
+  suffix: '',
+  includeThousandsSeparator: false,
+  allowDecimal: true,
+  decimalLimit: 2,
+  requireDecimal: false,
+  allowNegative: false,
+  allowLeadingZeroes: false,
+  integerLimit: 2
+});
+
+export const PHONE_MASK = createNumberMask({
+  prefix: '',
+  suffix: '',
+  includeThousandsSeparator: false,
+  allowDecimal: false,
+  decimalLimit: 0,
+  requireDecimal: false,
+  allowNegative: false,
+  allowLeadingZeroes: false,
+  integerLimit: 10
 });
