@@ -30,7 +30,7 @@ export class LoginComponent implements OnInit {
       ]),
       password: new FormControl('', [
         Validators.required,
-        Validators.minLength(10),
+        Validators.minLength(8),
         Validators.maxLength(16),
       ])
     }, Validators.required);
