@@ -1,14 +1,14 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { FormGroup, FormControl, Validators } from '@angular/forms';
-import { ActivatedRoute } from '@angular/router';
+import { ActivatedRoute, Router } from '@angular/router';
 import { Subscription } from 'rxjs';
 import { ProductsService } from 'src/app/modules/dashboard/pages/products/services/products.service';
-import { MatSnackBar } from '@angular/material';
+import { MatSnackBar } from '@angular/material/snack-bar';
 import { SnackbarComponent } from 'src/app/modules/shared/components/snackbar/snackbar.component';
 import { PRODUCTS_LANGUAGE } from 'src/app/modules/dashboard/pages/products/data/language';
 import { ACCOUNT_LANGUAGE } from 'src/app/modules/account/data/language';
 import { SNACKBAR_CONFIG } from 'src/app/modules/dashboard/pages/products/data/data';
-import { CURRENCY_MASK, NUMBER_MASK } from 'src/app/directives/currency-mask.directive';
+import { CURRENCY_MASK, NUMBER_MASK, PERCENTAGE_MASK } from 'src/app/directives/currency-mask.directive';
 
 @Component({
   selector: 'app-edit-product',
@@ -27,23 +27,29 @@ export class EditProductComponent implements OnInit, OnDestroy {
   private _subscriptionURL: Subscription;
   public lanProduct = PRODUCTS_LANGUAGE;
   public language = ACCOUNT_LANGUAGE;
+  public percentageMask = PERCENTAGE_MASK;
   public currencyMask = CURRENCY_MASK;
   public numberMask = NUMBER_MASK;
-  public showPricingPerDay = true;
   public imgChanged = false;
   public loading = false;
+  public permission: any;
 
   constructor(
+    private _router: Router,
     private _route: ActivatedRoute,
     private _snackBar: MatSnackBar,
     private _productService: ProductsService
   ) { }
 
   ngOnInit() {
+    this.permission = JSON.parse(localStorage.getItem('permission'));
     this.getProduct();
     this.editProductForm = new FormGroup({
       image: new FormControl(),
       name: new FormControl('', [
+        Validators.required,
+      ]),
+      sku: new FormControl('', [
         Validators.required,
       ]),
       brand: new FormControl('', [
@@ -67,41 +73,54 @@ export class EditProductComponent implements OnInit, OnDestroy {
       wholesaleQuantity: new FormControl('', [
         Validators.required,
       ]),
-      mondayPrice: new FormControl('', [
+      wholesalePriceG: new FormControl('', [
         Validators.required,
       ]),
-      tuesdayPrice: new FormControl('', [
+      wholesaleQuantityG: new FormControl('', [
         Validators.required,
       ]),
-      wednesdayPrice: new FormControl('', [
+      // mondayPrice: new FormControl('', [
+      //   Validators.required,
+      // ]),
+      // tuesdayPrice: new FormControl('', [
+      //   Validators.required,
+      // ]),
+      // wednesdayPrice: new FormControl('', [
+      //   Validators.required,
+      // ]),
+      // thursdayPrice: new FormControl('', [
+      //   Validators.required,
+      // ]),
+      // fridayPrice: new FormControl('', [
+      //   Validators.required,
+      // ]),
+      // saturdayPrice: new FormControl('', [
+      //   Validators.required,
+      // ]),
+      // sundayPrice: new FormControl('', [
+      //   Validators.required,
+      // ]),
+      sellerCommissionRetail: new FormControl('', [
         Validators.required,
       ]),
-      thursdayPrice: new FormControl('', [
+      sellerCommissionWholesale: new FormControl('', [
         Validators.required,
       ]),
-      fridayPrice: new FormControl('', [
-        Validators.required,
-      ]),
-      saturdayPrice: new FormControl('', [
-        Validators.required,
-      ]),
-      sundayPrice: new FormControl('', [
-        Validators.required,
-      ]),
-      sellerCommission: new FormControl('', [
+      sellerCommissionWholesaleG: new FormControl('', [
         Validators.required,
       ])
     });
   }
 
   public getProduct() {
-    this._subscription = this._route.params.subscribe(params  => {
+    this._subscription = this._route.params.subscribe(params => {
       this.id = params['id'];
       this._productService.getProduct(this.id).subscribe(
         res => {
           this.dataSource = res;
           this.editProductForm.get('image').patchValue(res.img_preview_url);
           this.editProductForm.get('name').patchValue(res.name);
+          this.editProductForm.get('sku').patchValue(res.sku);
           this.editProductForm.get('brand').patchValue(res.brand);
           this.editProductForm.get('unitsPackage').patchValue(res.units_package);
           this.editProductForm.get('category').patchValue(res.category);
@@ -109,14 +128,18 @@ export class EditProductComponent implements OnInit, OnDestroy {
           this.editProductForm.get('retailPrice').patchValue(res.retail_price);
           this.editProductForm.get('wholesalePrice').patchValue(res.wholesale_price);
           this.editProductForm.get('wholesaleQuantity').patchValue(res.wholesale_quantity);
-          this.editProductForm.get('mondayPrice').patchValue(res.monday_price);
-          this.editProductForm.get('tuesdayPrice').patchValue(res.tuesday_price);
-          this.editProductForm.get('wednesdayPrice').patchValue(res.wednesday_price);
-          this.editProductForm.get('thursdayPrice').patchValue(res.thursday_price);
-          this.editProductForm.get('fridayPrice').patchValue(res.friday_price);
-          this.editProductForm.get('saturdayPrice').patchValue(res.saturday_price);
-          this.editProductForm.get('sundayPrice').patchValue(res.sunday_price);
-          this.editProductForm.get('sellerCommission').patchValue(res.seller_commission);
+          this.editProductForm.get('wholesalePriceG').patchValue(res.wholesale_priceG);
+          this.editProductForm.get('wholesaleQuantityG').patchValue(res.wholesale_quantityG);
+          // this.editProductForm.get('mondayPrice').patchValue(res.monday_price);
+          // this.editProductForm.get('tuesdayPrice').patchValue(res.tuesday_price);
+          // this.editProductForm.get('wednesdayPrice').patchValue(res.wednesday_price);
+          // this.editProductForm.get('thursdayPrice').patchValue(res.thursday_price);
+          // this.editProductForm.get('fridayPrice').patchValue(res.friday_price);
+          // this.editProductForm.get('saturdayPrice').patchValue(res.saturday_price);
+          // this.editProductForm.get('sundayPrice').patchValue(res.sunday_price);
+          this.editProductForm.get('sellerCommissionRetail').patchValue(res.seller_commission_retail);
+          this.editProductForm.get('sellerCommissionWholesale').patchValue(res.seller_commission_wholesale);
+          this.editProductForm.get('sellerCommissionWholesaleG').patchValue(res.seller_commission_wholesaleG);
         });
     });
   }
@@ -128,13 +151,15 @@ export class EditProductComponent implements OnInit, OnDestroy {
         this.savePhoto().then(
           response => {
             this.editProductForm.get('image').patchValue(response);
-            this._productService.editProduct( this.editProductForm.value, this.id );
+            this._productService.editProduct(this.editProductForm.value, this.id);
             this.openSnackBar();
+            this._router.navigate(['/dashboard/products/view/' + this.id]);
           }, error => console.error(error)
         );
       } else {
-        this._productService.editProduct( this.editProductForm.value, this.id);
+        this._productService.editProduct(this.editProductForm.value, this.id);
         this.openSnackBar();
+        this._router.navigate(['/dashboard/products/view/' + this.id]);
       }
     }
   }
@@ -152,19 +177,19 @@ export class EditProductComponent implements OnInit, OnDestroy {
     this.base64textString = btoa(binaryString);
   }
 
-  public preview( files ) {
-    if ( files.length === 0 ) {
+  public preview(files) {
+    if (files.length === 0) {
       return;
     }
     const MIME_TYPE = files[0].type;
-    if (MIME_TYPE.match( /image\/*/ ) == null) {
+    if (MIME_TYPE.match(/image\/*/) == null) {
       // mat-error
       return;
     }
     const READER = new FileReader();
     this.imagePath = files;
-    READER.readAsDataURL( files[0] );
-    READER.onload = ( _event ) => {
+    READER.readAsDataURL(files[0]);
+    READER.onload = (_event) => {
       this.imgURL = READER.result;
     };
     this.imgChanged = true;
@@ -172,7 +197,7 @@ export class EditProductComponent implements OnInit, OnDestroy {
 
   public savePhoto() {
     return new Promise((resolve, reject) => {
-      this._productService.imageUpload( this.base64textString ).then(
+      this._productService.imageUpload(this.base64textString).then(
         response => {
           this._subscriptionURL = response.subscribe(
             url => resolve(url),
@@ -196,10 +221,10 @@ export class EditProductComponent implements OnInit, OnDestroy {
   }
 
   ngOnDestroy() {
-    if ( this._subscription ) {
+    if (this._subscription) {
       this._subscription.unsubscribe();
     }
-    if ( this._subscriptionURL ) {
+    if (this._subscriptionURL) {
       this._subscriptionURL.unsubscribe();
     }
   }
