@@ -1,7 +1,7 @@
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { Subscription } from 'rxjs';
-import { MatSnackBar } from '@angular/material';
-import { MatDialog } from '@angular/material';
+import { MatSnackBar } from '@angular/material/snack-bar';
+import { MatDialog } from '@angular/material/dialog';
 import { ActivatedRoute, Router } from '@angular/router';
 import { ProductsService } from 'src/app/modules/dashboard/pages/products/services/products.service';
 import { SnackbarComponent } from 'src/app/modules/shared/components/snackbar/snackbar.component';
@@ -19,17 +19,17 @@ export class ViewProductComponent implements OnInit, OnDestroy {
 
   public lanProduct = PRODUCTS_LANGUAGE;
   public language = ACCOUNT_LANGUAGE;
-  private _subscription: Subscription;
-  private _subscriptionService: Subscription;
   public id: string;
   public dataSource: any;
+  private _subscription: Subscription;
+  private _subscriptionService: Subscription;
 
   constructor(
-    private route: ActivatedRoute,
-    private router: Router,
-    private productService: ProductsService,
-    private snackBar: MatSnackBar,
-    private dialog: MatDialog,
+    private _route: ActivatedRoute,
+    private _router: Router,
+    private _productService: ProductsService,
+    private _snackBar: MatSnackBar,
+    private _dialog: MatDialog,
   ) { }
 
   ngOnInit() {
@@ -37,9 +37,9 @@ export class ViewProductComponent implements OnInit, OnDestroy {
   }
 
   public getProduct() {
-    this._subscription = this.route.params.subscribe(params  => {
+    this._subscription = this._route.params.subscribe(params => {
       this.id = params['id'];
-      this._subscriptionService = this.productService.getProduct(this.id).subscribe(
+      this._subscriptionService = this._productService.getProduct(this.id).subscribe(
         res => {
           this.dataSource = res;
         });
@@ -47,24 +47,39 @@ export class ViewProductComponent implements OnInit, OnDestroy {
   }
 
   public editProduct() {
-    this.router.navigate(['/dashboard/products/edit/' + this.id]);
+    this._router.navigate(['/dashboard/products/edit/' + this.id]);
+  }
+
+  public registerPromotion() {
+    this._router.navigate(['/dashboard/products/register-promotion/' + this.id]);
+  }
+
+  public editPromotion() {
+    this._router.navigate(['/dashboard/products/edit-promotion/' + this.id]);
+  }
+
+  public removeProduct() {
+    this._productService.removeProduct(this.id);
+    this._router.navigate(['/dashboard/products/']);
+    this.openSnackBarDeleted();
+  }
+
+  public removePromotion() {
+    this._productService.removePromotion(this.id);
+    this.openSnackBarPromoDeleted();
+    this._router.navigate(['/dashboard/products/view/' + this.id]);
   }
 
   public setDisponibility(event) {
     if (event.checked === true) {
-      this.productService.setDisponibility(this.id, event.checked);
+      this._productService.setDisponibility(this.id, event.checked);
     } else {
       this.openDialogDisable(event.checked);
     }
   }
 
-  public removeProduct() {
-    this.productService.removeProduct(this.id);
-    this.openSnackBarDeleted();
-  }
-
   public openSnackBarDeleted() {
-    this.snackBar.openFromComponent(SnackbarComponent, {
+    this._snackBar.openFromComponent(SnackbarComponent, {
       duration: SNACKBAR_CONFIG.duration,
       verticalPosition: SNACKBAR_CONFIG.verticalPosition,
       horizontalPosition: SNACKBAR_CONFIG.horizontalPosition,
@@ -74,8 +89,19 @@ export class ViewProductComponent implements OnInit, OnDestroy {
     });
   }
 
+  public openSnackBarPromoDeleted() {
+    this._snackBar.openFromComponent(SnackbarComponent, {
+      duration: SNACKBAR_CONFIG.duration,
+      verticalPosition: SNACKBAR_CONFIG.verticalPosition,
+      horizontalPosition: SNACKBAR_CONFIG.horizontalPosition,
+      data: {
+        text: PRODUCTS_LANGUAGE.snackbarPromoDeleted
+      }
+    });
+  }
+
   public openSnackBarDisabled() {
-    this.snackBar.openFromComponent(SnackbarComponent, {
+    this._snackBar.openFromComponent(SnackbarComponent, {
       duration: SNACKBAR_CONFIG.duration,
       verticalPosition: SNACKBAR_CONFIG.verticalPosition,
       horizontalPosition: SNACKBAR_CONFIG.horizontalPosition,
@@ -86,15 +112,17 @@ export class ViewProductComponent implements OnInit, OnDestroy {
   }
 
   public openDialogDisable(eventValue) {
-    const dialogRef = this.dialog.open(DialogComponent, {
-      data: {text: PRODUCTS_LANGUAGE.dialogDisable,
-             accept: true,
-             action: PRODUCTS_LANGUAGE.disable}
+    const dialogRef = this._dialog.open(DialogComponent, {
+      data: {
+        text: PRODUCTS_LANGUAGE.dialogDisable,
+        accept: true,
+        action: PRODUCTS_LANGUAGE.disable
+      }
     });
     dialogRef.afterClosed().subscribe(
       response => {
         if (response) {
-          this.productService.setDisponibility(this.id, eventValue);
+          this._productService.setDisponibility(this.id, eventValue);
           this.openSnackBarDisabled();
         }
       }
@@ -102,10 +130,12 @@ export class ViewProductComponent implements OnInit, OnDestroy {
   }
 
   public openDialogDelete() {
-    const dialogRef = this.dialog.open(DialogComponent, {
-      data: {text: PRODUCTS_LANGUAGE.dialogDelete,
-             accept: true,
-             action: PRODUCTS_LANGUAGE.remove}
+    const dialogRef = this._dialog.open(DialogComponent, {
+      data: {
+        text: PRODUCTS_LANGUAGE.dialogDelete,
+        accept: true,
+        action: PRODUCTS_LANGUAGE.remove
+      }
     });
     dialogRef.afterClosed().subscribe(
       response => {
@@ -116,13 +146,30 @@ export class ViewProductComponent implements OnInit, OnDestroy {
     );
   }
 
+  public openDialogDeletePromo() {
+    const dialogRef = this._dialog.open(DialogComponent, {
+      data: {
+        text: PRODUCTS_LANGUAGE.dialogDeletePromo,
+        accept: true,
+        action: PRODUCTS_LANGUAGE.delete
+      }
+    });
+    dialogRef.afterClosed().subscribe(
+      response => {
+        if (response) {
+          this.removePromotion();
+        }
+      }
+    );
+  }
+
   ngOnDestroy() {
-    if ( this._subscription ) {
+    if (this._subscription) {
       this._subscription.unsubscribe();
     }
-    if ( this._subscriptionService ) {
+    if (this._subscriptionService) {
       this._subscriptionService.unsubscribe();
     }
-   }
+  }
 
 }
