@@ -1,4 +1,4 @@
-import { MatSnackBarConfig } from '@angular/material/snack-bar'
+import { MatSnackBarConfig } from '@angular/material/snack-bar';
 
 export const SNACKBAR_CONFIG: MatSnackBarConfig = {
     duration: 5000,
