@@ -7,16 +7,21 @@ export const PRODUCTS_LANGUAGE = {
     disabled: 'Deshabilitado',
     disable: 'Deshabilitar',
     remove: 'Dar de baja',
+    removeProduct: 'Dar de baja producto',
     save: 'Guardar',
     edit: 'Editar',
     cancel: 'Cancelar',
+    delete: 'Eliminar',
+    search: 'Buscar...',
     snackbarAdd: 'Producto dado de alta exitosamente',
     snackbarDisabled: 'Producto deshabilitado',
     snackbarDeleted: 'Producto dado de baja exitosamente',
     snackbarEdited: 'Cambios guardados con éxito',
+    snackbarAddPromotion: 'Promoción creada',
+    snackbarPromoDeleted: 'Promoción eliminada exitosamente',
     dialogDelete: '¿Estás seguro que deseas dar de baja este producto?',
     dialogDisable: '¿Estás seguro que deseas deshabilitar este producto?',
-    createPromo: 'Crear promoción',
+    dialogDeletePromo: '¿Estás seguro que deseas eliminar esta promoción?',
     product: {
       name: 'Nombre',
       brand: 'Marca',
@@ -40,12 +45,15 @@ export const PRODUCTS_LANGUAGE = {
       retailPrice: 'Precio minorista:',
       wholesalePrice: 'Precio mayorista:',
       wholesaleQuantity: 'Cantidad para mayoreo:',
+      wholesalePriceG: 'Precio gran mayoreo:',
+      wholesaleQuantityG: 'Cantidad para gran mayoreo:',
       units: 'Piezas',
     },
     pricing: {
       pricing_by_day: 'Habilitar precios por día',
       currencySymbol: '$',
       country: 'MXN',
+      percentage: '%',
       monday: 'Lunes',
       tuesday: 'Martes',
       wednesday: 'Miércoles',
@@ -53,6 +61,16 @@ export const PRODUCTS_LANGUAGE = {
       friday: 'Viernes',
       saturday: 'Sábado',
       sunday: 'Domingo',
-      sellerCommission: 'Comisión a vendedor:'
+      sellerCommissionRetail: 'Comisión a vendedor por menudeo:',
+      sellerCommissionWholesale: 'Comisión a vendedor por mayoreo:',
+      sellerCommissionWholesaleG: 'Comisión a vendedor por gran mayoreo:'
+    },
+    promotion: {
+      promotion: 'Promoción',
+      editPromotion: 'Editar promoción',
+      deletePromotion: 'Eliminar promoción',
+      createPromotion: 'Crear promoción',
+      startDate: 'Fecha de inicio:',
+      endDate: 'Fecha límite:',
     }
   };
