@@ -1,41 +1,61 @@
-import { Component, OnInit, OnDestroy } from '@angular/core';
+import { Component, OnInit, OnDestroy, ViewChild, AfterViewInit } from '@angular/core';
 import { ProductsService } from 'src/app/modules/dashboard/pages/products/services/products.service';
 import { Subscription } from 'rxjs';
-import { Router } from '@angular/router';
+import { MatTableDataSource } from '@angular/material/table';
+import { MatPaginator } from '@angular/material/paginator';
 import { PRODUCTS_LANGUAGE } from 'src/app/modules/dashboard/pages/products/data/language';
+import { PAGINATION } from 'src/app/modules/shared/components/paginator/data/data';
 
 @Component({
   selector: 'app-products',
   templateUrl: './products.component.html',
   styleUrls: ['./products.component.scss']
 })
-export class ProductsComponent implements OnInit, OnDestroy {
+export class ProductsComponent implements OnInit, OnDestroy, AfterViewInit {
 
   public language = PRODUCTS_LANGUAGE;
   public displayedColumns: string[] = ['img', 'name', 'has_promo', 'brand', 'content', 'quantity', 'retailPrice', 'wholesalePrice'];
-  public dataSource: any;
-  public products;
+  public dataSource = new MatTableDataSource();
+  public indexProducts = 0;
   public subscriptionProducts: Subscription;
+  public pagination = PAGINATION;
+  @ViewChild(MatPaginator) paginator: MatPaginator;
 
   constructor(
     private productsService: ProductsService,
-    private router: Router
   ) { }
 
   ngOnInit() {
     this.getProducts();
   }
 
+  ngAfterViewInit() {
+    this.dataSource.paginator = this.paginator;
+  }
+
   public getProducts() {
     this.subscriptionProducts = this.productsService.getAllProducts().subscribe(
       res => {
-        this.dataSource = res;
+        this.dataSource.data = res;
+        const data = [];
+        for (let i = 0; i < this.pagination.perPage; i++) {
+          if (res[i]) {
+            data.push(res[i]);
+            this.indexProducts = this.indexProducts + i;
+          }
+        }
       }
     );
   }
 
+  public doFilter = (value: string) => {
+    this.dataSource.filter = value.trim().toLocaleLowerCase();
+  }
+
   ngOnDestroy() {
-    this.subscriptionProducts.unsubscribe();
+    if (this.subscriptionProducts) {
+      this.subscriptionProducts.unsubscribe();
+    }
   }
 
 }
