@@ -1,12 +1,18 @@
+import { MatSnackBarConfig } from '@angular/material/snack-bar';
+
 export const ROLES = {
-    STORER: 2,
-    VENDOR: 1,
-    ADMIN: 0
+    VENDOR: 0,
+    STORER: 1,
+    ADMIN: 2
 };
 
 export const PERMISIONS = {
     USER_REGISTRATION: 0,
-    PRICE_EDITION: 1,
-    CREATE_EDIT_PROMOTIONS: 2
+    PRICE_EDITION: 1
 };
 
+export const SNACKBAR_CONFIG: MatSnackBarConfig = {
+    duration: 5000,
+    horizontalPosition: 'center',
+    verticalPosition: 'bottom'
+};
