@@ -4,18 +4,30 @@ export const USERS_LANGUAGE = {
     storer: 'Almacenista',
     vendor: 'Vendedor',
     admin: 'Administrador',
+    search: 'Buscar...',
+    cancel: 'Cancelar',
+    delete: 'Eliminar',
+    disable: 'Deshabilitar',
+    dialogDelete: '¿Estás seguro que deseas eliminar a este usuario?',
+    dialogDisable: '¿Estás seguro que deseas deshabilitar a este usuario?',
+    snackbarDeleted: 'Usuario eliminado exitosamente',
+    snackbarDisable: 'Usuario deshabilitado exitosamente',
     user: {
         name: 'Nombre',
+        username: 'Nombre de Usuario',
         rol: 'Rol',
+        route: 'Ruta',
         dateCreated: 'Creado',
-        lastConexion: 'Última conexión',
+        lastConexion: 'Última conexión:',
         status: 'Estatus',
+        notValid: 'No aplica'
     },
     newUser: {
         title: 'Nuevo Usuario',
         typeRol: 'Tipo de rol',
         route: 'Ruta',
         name: 'Nombre de la persona',
+        username: 'Nombre de usuario',
         email: 'Correo electrónico'
     },
     actions: 'Acciones',
@@ -23,5 +35,20 @@ export const USERS_LANGUAGE = {
         userRegistration: 'Alta de usuarios',
         priceEdition: 'Edición de precios',
         createEditPromotions: 'Creación y edición de promociones'
-    }
+    },
+    edit: {
+        created: 'Creado:',
+        editUser: 'Editar usuario',
+        name: 'Nombre:',
+        password: 'Contraseña:',
+        route: 'Ruta:',
+        actions: 'Acciones:',
+        status: 'Estatus:',
+        enabled: 'Habilitado',
+        disabled: 'Deshabilitado',
+        deleteUser: 'Eliminar usuario',
+        saveUser: 'Guardar usuario',
+        userRegistration: 'Alta de usuarios',
+        editPrices: 'Edición de precios'
+    },
 };
