@@ -1,5 +1,6 @@
 export class RouteModel {
-    id: string;
-    name: string;
+    id?: string;
+    name?: string;
     owner?: string;
+    seller?: string;
 }
