@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Inject } from '@angular/core';
-import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material';
+import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
 import { PRODUCTS_LANGUAGE } from 'src/app/modules/dashboard/pages/products/data/language';
 
 @Component({
@@ -21,7 +21,7 @@ export class DialogComponent implements OnInit {
   ngOnInit() {
   }
 
-  public removeProduct () {
+  public removeProduct() {
   }
 
   public accept() {
