@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Inject } from '@angular/core';
-import { MAT_SNACK_BAR_DATA, MatSnackBarRef } from '@angular/material';
+import { MAT_SNACK_BAR_DATA, MatSnackBarRef } from '@angular/material/snack-bar';
 
 @Component({
   selector: 'app-snackbar',
@@ -13,7 +13,7 @@ export class SnackbarComponent implements OnInit {
     @Inject(MAT_SNACK_BAR_DATA)
     public data: any,
     private snackBarRef: MatSnackBarRef<SnackbarComponent>
-  ) {}
+  ) { }
 
   ngOnInit() {
   }
