@@ -1,38 +1,44 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
+import { CommonModule, DatePipe } from '@angular/common';
 import { RouterModule } from '@angular/router';
 import { HttpClientModule } from '@angular/common/http';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { TextMaskModule } from 'angular2-text-mask';
+import { AgmCoreModule } from '@agm/core';
+import { AgmSnazzyInfoWindowModule } from '@agm/snazzy-info-window';
+import { MomentModule } from 'ngx-moment';
+import { ChartsModule } from 'ng2-charts';
+
 import 'hammerjs';
 
-
 /* ANGULAR MATERIAL */
-import {
-  MatButtonModule,
-  MatInputModule,
-  MatIconModule,
-  MatTableModule,
-  MatToolbarModule,
-  MatSidenavModule,
-  MatListModule,
-  MatMenuModule,
-  MatSelectModule,
-  MatRadioModule,
-  MatSlideToggleModule,
-  MatCheckboxModule,
-  MatCardModule,
-  MatSnackBarModule,
-  MatDialogModule,
-  MatProgressBarModule,
-  MatDatepickerModule,
-  MatNativeDateModule
-} from '@angular/material';
+import { MatButtonModule } from '@angular/material/button';
+import { MatInputModule } from '@angular/material/input';
+import { MatIconModule } from '@angular/material/icon';
+import { MatTableModule } from '@angular/material/table';
+import { MatToolbarModule } from '@angular/material/toolbar';
+import { MatSidenavModule } from '@angular/material/sidenav';
+import { MatListModule } from '@angular/material/list';
+import { MatMenuModule } from '@angular/material/menu';
+import { MatSelectModule } from '@angular/material/select';
+import { MatRadioModule } from '@angular/material/radio';
+import { MatSlideToggleModule } from '@angular/material/slide-toggle';
+import { MatCheckboxModule } from '@angular/material/checkbox';
+import { MatCardModule } from '@angular/material/card';
+import { MatSnackBarModule } from '@angular/material/snack-bar';
+import { MatDialogModule } from '@angular/material/dialog';
+import { MatProgressBarModule } from '@angular/material/progress-bar';
+import { MatDatepickerModule } from '@angular/material/datepicker';
+import { MatNativeDateModule } from '@angular/material/core';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatPaginatorModule, MatPaginatorIntl } from '@angular/material/paginator';
+
 import { LayoutModule } from '@angular/cdk/layout';
+import { SatDatepickerModule, SatNativeDateModule } from 'saturn-datepicker';
 
 /* FIREBASE */
 import { AngularFireModule } from '@angular/fire';
-import { AngularFirestoreModule, FirestoreSettingsToken } from '@angular/fire/firestore';
+import { AngularFirestoreModule, SETTINGS } from '@angular/fire/firestore';
 import { AngularFireStorageModule } from '@angular/fire/storage';
 import { AngularFireAuthModule } from '@angular/fire/auth';
 import { AngularFireDatabase } from '@angular/fire/database';
@@ -49,7 +55,7 @@ import { ToastComponent } from '../dashboard/components/toast/toast.component';
     PaginatorComponent,
     SnackbarComponent,
     DialogComponent,
-    ToastComponent
+    ToastComponent,
   ],
   imports: [
     CommonModule,
@@ -80,7 +86,17 @@ import { ToastComponent } from '../dashboard/components/toast/toast.component';
     MatDialogModule,
     MatProgressBarModule,
     MatDatepickerModule,
-    MatNativeDateModule
+    MatNativeDateModule,
+    MatFormFieldModule,
+    MatPaginatorModule,
+    AgmCoreModule.forRoot({
+      apiKey: 'AIzaSyAskN-VARhdUaLDaeA9chNu4vpveNjNs4s'
+    }),
+    AgmSnazzyInfoWindowModule,
+    SatDatepickerModule,
+    SatNativeDateModule,
+    MomentModule,
+    ChartsModule
   ],
   exports: [
     AngularFirestoreModule,
@@ -110,13 +126,26 @@ import { ToastComponent } from '../dashboard/components/toast/toast.component';
     ToastComponent,
     MatProgressBarModule,
     MatDatepickerModule,
+    MatFormFieldModule,
+    MatPaginatorModule,
+    AgmCoreModule,
+    AgmSnazzyInfoWindowModule,
+    SatDatepickerModule,
+    SatNativeDateModule,
+    MomentModule,
+    ChartsModule
   ],
   providers: [
     AngularFireDatabase,
-    { provide: FirestoreSettingsToken, useValue: {} },
-    MatDatepickerModule
+    { provide: SETTINGS, useValue: {} },
+    MatDatepickerModule,
+    DatePipe,
+    { provide: MatPaginatorIntl, useClass: PaginatorComponent }
   ],
-  entryComponents: [SnackbarComponent,
-                    DialogComponent, ToastComponent]
+  entryComponents: [
+    SnackbarComponent,
+    DialogComponent,
+    ToastComponent
+  ]
 })
 export class SharedModule { }
